+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - 修复 TimePicker 图标样式被默认图标样式覆盖
 - 修复 在 mounted 钩子函数中改变 Select 绑定值不生效的问题
 - 修复 在多个依次出现的 Dialog 或 Message Box 全部关闭后页面有几率不可滚动的问题
+- 新增 时间、日期选择器增加 align 属性，可设置对齐方式
+- 新增 TableColumn 的 align 属性
 
 #### 非兼容性更新
 - Select 组件样式的 `display` 属性默认值修改为 `block`